It looks like you've made some changes to your `commit.sh` script!
Here's a breakdown of what I see:

* You've updated the `commit_msg` variable assignment by modifying the prompt passed to the `ollama run` command.
* The new prompt includes an additional instruction, "Your response should only contain the message and nothing else:", which is likely intended to guide the LLaMA model's output.

With this change, when you run the script, it will generate a commit message using the LLaMA model with a more specific prompt. The resulting commit message will be used for your Git commit.

Is there anything in particular you'd like me to help with regarding this updated script?